@@ -1,11 +1,15 @@
 module github.com/jmpa-io/up-go
 
-go 1.20
+go 1.22.0
 
-require github.com/rs/zerolog v1.30.0
+toolchain go1.23.4
+
+require go.opentelemetry.io/otel v1.33.0
 
 require (
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.19 // indirect
-	golang.org/x/sys v0.11.0 // indirect
+	github.com/go-logr/logr v1.4.2 // indirect
+	github.com/go-logr/stdr v1.2.2 // indirect
+	go.opentelemetry.io/auto/sdk v1.1.0 // indirect
+	go.opentelemetry.io/otel/metric v1.33.0 // indirect
+	go.opentelemetry.io/otel/trace v1.33.0 // indirect
 )
