@@ -1,7 +1,9 @@
 package main
 
 import (
+	"context"
 	"fmt"
+	"log/slog"
 	"os"
 
 	"github.com/jmpa-io/up-go"
@@ -9,8 +11,13 @@ import (
 
 func main() {
 
+	ctx := context.TODO()
+
+	// retrieve token.
+	token := os.Getenv("UP_TOKEN")
+
 	// setup client.
-	c, err := up.New("xxxx")
+	c, err := up.New(ctx, token, up.WithLogLevel(slog.LevelWarn))
 	if err != nil {
 		fmt.Printf("failed to setup client: %v\n", err)
 		os.Exit(1)
@@ -18,10 +25,10 @@ func main() {
 
 	// do something with the client..
 	// like send a ping to the api to check if the token is valid.
-	p, err := c.Ping()
+	p, err := c.Ping(ctx)
 	if err != nil {
 		fmt.Printf("failed to ping: %v\n", err)
 		os.Exit(1)
 	}
-	fmt.Printf("%+v\n", p)
+	fmt.Printf("%s\n", p.Meta.StatusEmoji)
 }
