@@ -1,27 +1,35 @@
 package up
 
-import "net/http"
+import (
+	"context"
+	"net/http"
+
+	"go.opentelemetry.io/otel"
+)
 
 // Ping represents a ping event sent from the API.
 type Ping struct {
-	Id          string `json:"id"`
-	StatusEmoji string `json:"statusEmoji"`
-}
-
-// PingResponse represents the response returned from the API when
-// trying to ping the API.
-type PingResponse struct {
-	Meta Ping `json:"meta"`
+	Meta struct {
+		Id          string `json:"id"`
+		StatusEmoji string `json:"statusEmoji"`
+	} `json:"meta"`
 }
 
 // Ping makes a ping request to the API.
 // https://developer.up.com.au/#get_util_ping
-func (c *Client) Ping() (*Ping, error) {
-	cr := clientRequest{
+func (c *Client) Ping(ctx context.Context) (*Ping, error) {
+
+	// setup tracing.
+	newCtx, span := otel.Tracer(c.tracerName).Start(ctx, "Ping")
+	defer span.End()
+
+	// ping.
+	var p *Ping
+	if _, err := c.sender(newCtx, senderRequest{
 		method: http.MethodGet,
 		path:   "/util/ping",
+	}, &p); err != nil {
+		return nil, err
 	}
-	var resp *PingResponse
-	_, err := c.sender(cr, &resp)
-	return &resp.Meta, err
+	return p, nil
 }
